remove long that prevents int overflow
the restriction placed on n and m is: m*n < int.MaxValue. This will allow to store TP with cost table of 16GB, which seems enough

+++ TpSolver.Tests/AllocationMatrixTests.cs
@@ -39,7 +39,7 @@ public class AllocationMatrixTests
                 { 0, 4, 0, 5 },
             }
         );
-        long bc = am.CountBasic();
+        int bc = am.CountBasic();
         Assert.Equal(5, bc);
 
         am[0, 3] = am[0, 3].AsBasic();
